+++ tests/test_sources.py
@@ -1,3 +1,5 @@
+import pathlib
+
 import hypothesis
 import numpy as np
 from hypothesis import given, strategies as st
@@ -16,7 +18,11 @@ from pyechelle.simulator import check_url_exists
 def test_sources(wl, bw, source_name):
     wavelength = np.linspace(wl, wl + bw, 1000, dtype=np.float)
     print(f"Test {source_name}...")
-    source = getattr(sources, source_name)()
+    if source_name == 'CSV':
+        source = getattr(sources, source_name)(pathlib.Path(__file__).parent.joinpath('test_data/test_eff.csv'),
+                                               wavelength_unit='micron', stellar_target=False)
+    else:
+        source = getattr(sources, source_name)()
     sd = source.get_spectral_density(wavelength)
     assert isinstance(sd, tuple) or isinstance(sd, np.ndarray)